requirements.txt
aiogram
mistralai
asyncpg
Pillow
python-dotenv
APScheduler
aiofiles
psycopg2-binary